fix: Documentation warnings and unnecessary cast in tests
- Fix rustdoc broken intra-doc links by escaping square brackets in apn_dnn.rs
- Remove unnecessary u16 cast in tests/messages.rs to resolve clippy warning
- Both changes maintain existing functionality while eliminating CI warnings

+++ tests/messages.rs
@@ -358,7 +358,7 @@ fn test_session_modification_response_marshal_unmarshal() {
         0x112233,
     );
     let payload_len = cause_ie.len() + pdr_ie.len();
-    header.length = (payload_len + header.len() - 4) as u16;
+    header.length = payload_len + header.len() - 4;
 
     let res = SessionModificationResponse {
         header,
